requirements.txt
bitarray==3.5.0
bitstring==4.3.1
cffi==1.17.1
click==8.1.8
cryptography==45.0.5
esptool==5.0.0
intelhex==2.3.0
markdown-it-py==3.0.0
mdurl==0.1.2
pycparser==2.22
Pygments==2.19.2
pyserial==3.5
PyYAML==6.0.2
reedsolo==1.7.0
rich==14.0.0
rich-click==1.8.9
typing_extensions==4.14.1
